backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3


